cookbook/pyproject.toml
[tool.poetry]
name = "gram_server"
version = "0.1.0"
description = ""
authors = ["Dhruval Patel <dhruval@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
django-mptt = "0.15.0"
graphene-django = "^2.15.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.4.5"
structlog = "^22.1.0"
boto3 = "1.28.2"
botocore = "1.31.80"
pdfcrowd = "^5.2.2"
beautifulsoup4 = "^4.10.0"
django-filter = "^22.1"
pandas = "1.5.0"
django-jsonfield = "^1.4.1"
openpyxl = "3.0.7"
html5lib = "^1.1"
Pillow = "^9.2.0"
uvicorn = "^0.18.2"
django-clone = "^3.0.3"
twilio = "^7.12.1"
django-graphql-jwt = "^0.3.4"
celery = {extras = ["sqs"], version = "^5.2.7"}
cognitojwt = "^1.4.1"
django-storages = "^1.13.1"
django-celery-results = "^2.4.0"
django-prices = "^2.3.0"
uvloop = "^0.17.0"
httptools = "^0.5.0"
django-countries = "^7.5.1"
graphql-ws = "^0.4.4"
channels = "3.0.4"
channels-redis = "^4.1.0"
websockets = "^11.0.3"
django-redis = "^5.3.0"
geopy = "^2.3.0"
django-celery-beat = "^2.5.0"
docutils = "^0.20.1"
moto = "^3.1.16"
pymssql = "^2.2.8"
aiodataloader = "^0.4.0"

[tool.poetry.group.dev.dependencies]
django-cors-headers = "^3.7.0"
pre-commit = "^2.13.0"
pytest-django = "^4.4.0"
pytest-mock = "^3.6.1"
snapshottest = "^0.6.0"
pytest-cov = "^2.12.1"
django-environ = "^0.4.5"
factory-boy = "^3.2.0"
pytest-freezegun = "^0.4.2"
pytest-network = "^0.0.1"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
